[FIX] barcodes: set_focus on last scanned element
When we scan a element or update its quantity, we don't always
see the added or modified element on the screen, id we are in mobile
view and the element is not displaied on screen.

So to fix this, we are making a scrollTop to the element that have
been added or updated. We did it by adding to the 'KanbanView' method
'reload_record' a call to 'scrollTop' on the reloaded record. And calling
'scrollTop' on the last element of the list when we have to add a new
element to the kanban view.