[FIX] module: Test logger logs traceback line by line
This is not an issue when logging to a file or stdout, but when using
`--log-db` we got an `ir.logging` entry for each line of a traceback.

I had a chat with @antonylesuisse on this topic and we decided to remove
the backtick+space at the beginning of each traceback lines.

For the record or in case the backticks should be restored, I originally
came with this solution:

```diff
diff --git a/odoo/modules/module.py b/odoo/modules/module.py
index e6194a5fb10..1cb7d37b582 100644
--- a/odoo/modules/module.py
+++ b/odoo/modules/module.py
@@ -458,13 +458,8 @@ class TestStream(object):
     def write(self, s):
         if self.r.match(s):
             return
-        first = True
         level = logging.ERROR if s.startswith(('ERROR', 'FAIL', 'Traceback')) else logging.INFO
-        for c in s.splitlines():
-            if not first:
-                c = '` ' + c
-            first = False
-            self.logger.log(level, c)
+        self.logger.log(level, s.replace('\n', '\n` '))
```

which have the same formatting behaviour but outputs a single entry
instead of multiple logging entries.