[IMP] test_mail: correctly set aliases when performing tests
In order to really test a production environment we have to ensure
some mail related config parameters are set. This has an impact on
some tests.

Which is strange is that it lessens query number of a test that
already uses those parameters. Maybe XMas spirit.

Backport of 5c3e0391acda99c50bf1db4f46022726bfc739ac .