[FIX] hr_holidays: Add missing rule for multi company leaves
Purpose
=======

There's a ir.rule to retrict the access to leave types for a leave officer.

As there's no restriction on leaves, this could lead to access right errors when playing on the reporting views.

Example: Remove a filter on a grouped view. We want to group by leave types the remaining leaves --> Error.

Specification
=============

Restricting the leaves on others companies than the company the user is logged in should solve the issue.