[FIX] purchase: _anglo_saxon_purchase_move_lines
If you receive a PO and return it, both the receipt and returned move
have the same purchase_line_id.

In v11, we value the returned move not at the original price but at the
standard/average price of the moment or to a next price in the fifo
stack.

If you receive 10@10, 10@12 and return the second receipt, the return
will be valued at 10, it is therefore important for the anglo saxon move
lines to get the correct price on the moves to be able to post a correct
entry in the price difference account when creating a credit note.

The issue here is that the code was not adapted from v10 and when
getting the price on the move, it'll consider all the moves of the PO:
the receipt and the return. The fix is to only consider returned move
when getting the price on the move.

FIXME: return of return?
opw-1819353