[REF] mrp: update quantity to produce
`_generate_finished_moves` was adapted since it was called when a new
finished move needed to be created, however the move was confirmed with
the total quantity to produce before its product_qty was written on with
the difference quantity. This was wrong since the push rules were
created for the total quantity.

m