[ADD] project_expense: select project on expense directly
This commit provides a new bridge module, between project and expense. Its
purpose is to select directly a project on an expense. Indeed, for normal
user, it is not easy to know the analytic account on which the expense should
be reinvoiced. Selecting the project, will set the analytic account and make
this use case simplier to use day-to-day.
Project field on expense is only visible for project user, and selectable
projects are only the ones with an analytic account set.
This does not solve the reinvoice flow (still automatic), but it's a start.

Task-1911581