[ADD] sale_stock, sale_mrp: allow the user to cancel confirm sale order
Be able to cancel an order if the products have been returned.

Purpose this task, cancel the sale order if all the delivered products have been
returned.

After this commit, if all the products which we delivered to the customer, and
customer returned all the products, then user is able to cancel the sale order.

task-1919761

Closes #29920