[IMP] point_of_sale: Update time on IoTBox boot
Displays connected to an IoTBox should be able to show any URL.
Since we didn't have time synchronization, the time of the IoTBox was frequently a few weeks or months behind the real time.
SSL certificate created after the current time of the IotBox were rejected, making it impossible to access some URLs.
To fix this, we update the time everytime the Raspberry Pi boots.