[IMP] website_slides : refactor share modal to be used in other depending modules
This commit prepares the next one. It creates a generic template based on the slide share modal
in order to make this modal available for other modules depending of website.

In order to use the specific behaviour of slide_share, the class of share buttons
is given by the caller. So the template can be used just to share without doing anything
specific in other modules but can be still extended using JS class selector.

Task ID : 1950396
PR #31776