[REF] hw_escpos: Integrate hw_escpos with IoTBox's PrinterDriver
ESCPOS printer are now supported just like any other printer as they print PDFs.
This means that we can print any report with a receipt printer and that receipts
can be printed on any printer.

`hw_escpos` is left untouched to allow backward compatibility and should be
removed when v12 is no longer supported.

TaskID: 1939455