[IMP] crm: merge lead and opportunity form views
When redirecting to record, the wrong 'form' view will sometimes be
loaded. A 'hack' has been done in crm_lead.py to redirect the user to
the relevant 'form' view based on the value of the 'type' field.

However, this solution is not clean as it does not solve all cases
(multiple crm.lead records loaded from the activity systray, reloading
the page after redirection, etc.)

Thus, The goal of this task is to merge the crm.lead.form.lead and
crm.lead.form.opportunity formviews.

That's why merged lead and opportunity form views and adapt changes
based on that.

Note: Here some fields used as a twice in the form view just to stay
close to the current design. Still few things which are not possible
here for 'lead' and 'opportunity', that is
1. A dynamic 'placeholder' on 'name' field.
2. A 'domain' on 'partner_id' field.
3. A 'widget' on 'probability' field.

task-47118