[IMP] base: parent and child partners should have the same company
purpose of the commit is When creating a new child system and
one of them is from a different company and tries to log a message on all
But since one of them is from a different company so the record rule
triggers an error.

so if parent is set on contact then the company_id should be readonly and
it will automaitcally set the parent's company_id.
if parent's company_id changed then also change the company_id of all childs

Added the test cases for child/parent company so we know it works rebust without
the UI as well.

Task-1934813
closes: #32405