[IMP] web: form: buttons and statusbar: no more overflow
This rev. improves the way the content of the <header> tag in form
views is rendered, such that the statusbar buttons are always
displayed in a single line. The buttons take as much space they
need, and the statusbar take the remaining space. It computes how
many buttons it can display, and if there is not enough space for
all buttons, the remaining ones are automatically put in a 'More'
dropdown.

Task 1932190
Closes #33614

Co-authored-by: Parth Choksi <pch@odoo.com>