[IMP] point_of_sale: Add keysym mapping
Add keysym mapping to the IoTBox config to be used in KeyboardUSBDriver.

The Keysym -> Character mapping provided by Xlib only supports latin-1
encoded characters. For keysyms corresponding to special characters, X11
won't return anything. We provide a Keysym to Unicode mapping for most
of the characters not supported by X11.