[IMP] base,product,website_sale: move image fallback mixin -> variant
Backport of d3d58b60dc8c67d0b9926821ca13f0e6a3832a1d

Only the variant is using this code, so it was a bad idea to make the mixin
more complicated than it had to be.

Update the image tests to make sure this is working correctly. The test were
only testing the size of the images, now they also test the actual content.

Part of task-1949729

Original PR: odoo/odoo#32253