[IMP] website, base: make the website picture translatable
With this commit images are translatable now. To do this we have made image
src attribute translatable. In simple word, we will use different images for
different languages. To set the translatable image user will need to click on
the image in the translatable mode which will open the attribute translate
dialog where user can change the image.

task-33372