[REF] mail: remove private discussion support in mail gateway
Private discussion support in mail gateway means having incoming messages
not linked to a document. It means having a void model and res_id on related
messages.

Old private conversations in Odoo were implemented using messages not linked
to any document. It was possible to send an email to some people through the
UX and they were able to answer it. Messages were not attached to any document
and model.

It is not supported anymore since discussion between users is now done using
chat. Contacting partners / customers can be done on business documents
or on channels. Simple email conversation is not supported in Odoo as it
is not the purpose of the mailgateway.

This commit removes code related to private discussion in mail gateway and
some specific support in message_post method.

Related to task 1893155
Linked to PR #33340