[IMP] point_of_sale: cash change shouldn't create a payment
Before this commit, if a user creates PoS order of total amount $80,
Encode a cash payment of $100 and change to give back is $20.
journal entry creates two lines for both $20 and $100
it should create an only single line of $80, for the cash transaction.
this commit will merge change if it is cash method.

Task-1944130