[IMP] product, stock_account: track the product cost on product and template
Purpose of this commit is to track the cost price on product.

so we enabled tracking on field standard_price in product.product

but in product.template, field 'standard_price' is of computed type
so for product.template we can not put the tracking on that field and even template
do not have price.

so for Product Template when the 'standard price' is changed via "Update Cost"
button and if the product has only one variant then we post the message
forcefully in product.template which will be visible to the specific company only

Task-1903115