[FIX] survey : externalise survey web assets to avoid multiple load at website install
Web assets should not be loaded in the survey assets to display survey page
when the website is installed as the website already load those assets.

So thoses web assets are now externalised in another template that is loaded
only if survey is installed without website installed.

Task ID : 1930132
PR #32419