[IMP] project: limit change when linkin sub task to parent
When a task is linked o a perent task, some of its fields are altered to match the paren'ts corresponding fields. This might not be what the user wants when, for example he already has created that task, for a certain customer, the links it to a parent afterward. If he need to keep the original customer, he'll have to modify it by hand. And since the change is implicit, the user might not even realize what changes have taken place.

Though this is handy when quickly creating subtasks that should effectively use the parent's field.

We compromise by only updating subtask's fields with the value of the parent's, when the subtask didn't already define those values.

related task #1982072