[REF] product: remove product.price.history table
`get_history_price` was only used to compute the stock value at the
dote, but we now always work with all the stock.valuation.layers records
since the creation of the database to a date passed in the context.

if stock_account is installed then a layer will be created for each
standard price update.

task-1875873