[REF] website_slides: Remove slide.category
This commit removes the slide.category model for
UX purpose in the backend.

Instead of having a fully fledged slide.category model,
slide.slide will serve that purpose with a is_category flag.
This will allow to drag and drop slides and sections
in the channel form view

This change had an impact on the way slides were added/sorted
on the front-end. In fact, whenever a slide is added
from the front-end, a resequencing of all the slides in the course
has to be triggered.