[IMP] mail, *: adding database name in the url of company logo
* = auth_signup, digest, event, gamification, lunch, mass_mailing_crm,
mass_mailing_sale, portal, survey, website_crm_partner_assign,
website_mail_channel, website_profile, website_sale

In the mail templates, when the logo of the company is needed, by default,
we use the route <domain_name>/logo.

In some cases the system does not manage to "guess" which DB or company to use
and this gives the default Odoo logo which can be problematic.

In this commit, we're adding database name as a parameter in logo url to
overcome this issue.

task-1889370
closes #31772