[REF] sms: perform some guidelining and rename composer model
Purpose is to prepare future improvements in SMS module. In this commit we
apply some guidelines on xml ids and file naming, just a few improvements.

More importantly sms.send_sms model is renamed to sms.composer to avoid
underscores and to have a better name.

Related to task 1922163
Linked to PR #33510