[FIX] mail: fix negative operator issue on search of follower channel/partner
Before this commit:
  There was a generic issue that on custom filter trying to apply
filter by 'follower(channels)', ' follower(partner)' with negative
oprator (like: 'doesn't contain', 'is not equal to', 'is not set'
etc..) were not working because it was not handled by _search function.

After this commit:
  IF user search by the negative operator then convert negative operator
into the posistive operator and then find the follower and then use the
'not in' in main domain of many2many table.

task-1971489
Closes: #34101