[IMP] phone_validation, sms: improve phone validation formatting and add it in SMS dependencies
Purpose of this commit is to support more formatting options in phone
validation tools.

Indeed in addition to national and international formatting phonenumbers
library also supports E164 (international without spaces) and RFC3966
(beautification of phone numbers). Let us support them and ease the use
of our small tool methods in various addons, notably sms.

Another purpose of this commit is to add phone_validation in dependencies
of SMS application.

Phone_validation module adds tools to parse and format phone numbers. Adding
it in dependencies of SMS does not change anything functionally as

  * library is still optional. If phonenumbers library is not installed the
    features are skipped without crashing;
  * no model is modified as it introduces only tools and a mixin. Models are
    updated in crm_phone_validation that updates lead and partner models.
    However it is a separated module that is not automatically installed.
    There is therefore no harm adding phone_validation as dependency for SMS;

Adding it allows to use tools method instead of having to manually redo
the whole parsing and formatting a bit differently in SMS notably.

Related to task 1922163
Linked to PR #33510