[IMP] sale: set company required
The sales order is a important model. Since it generates invoices and other business
documents, the company_id field should be required as the company must be propagate
to generated documents.
Moreover, a sales order without a company raises a error, and we need a company to
compute the multi currency rate.

This commit fixes that bug, by setting the company_id field required on sales order
and sale order line.

Task-1999686