[IMP] website: render available fonts
Currently if a theme has more fonts that the default number (6), it
overrides the customize dialog. Now, the number of fonts to add in the
list is found automatically and fonts are rendered automatically.

To do so, instead of defining a <selection/> tag with 6+ <opt/> tag, we
are now able to define an unique new tag: <fontselection/> which will do
automatically the job.

task-1974659