[IMP] website_sale_slides: improve course purchase flow V2
Before this task, when a customer bought a course, he was directly
redirect to the checkout process.
After this task, the customer can either Add to Cart or Buy Now
(if the Buy Now option is activated)

[IMP] website_sale_product_configurator: delete regex for redirection

When a product has an optional product, you have the possibility to add
in your cart the optional product thanks to a modal. Before, the button
"Add to Cart" was only available in the shop (/shop/), but now it is also available
on a course (/slides/). So to be redirected on the cart when we click on the button
"Proceed to Checkout" for a course, we needed to apply the redirection also
for the course.

Closes #34224