[IMP] web: Add a custom lib to handle multirange input
Range input with multiple values are not native in html-5 nor Bootstrap.
The custom lib "multirange" is inspired from this polyfill:
https://github.com/leaverou/multirange

The multirange will display the two values as one range input with two
cursor linked with a background. This is compatible with Bootstrap and
the theme colors if the "custom-range" class is added.

There is 2 number inputs on the right and left of the multirange to
display and allow quick and precise value modifications. They are
initialised with the same value provided to the input (min, max, step)

There is 2 events that are added to the input:
    - oldRangeValue: Trigger when the user click on a cursor or on focus
    of the right or left number input.

    - newRangeValue: Trigger when the user release a cursor or on focus
    out of the right or left number input.

The options available for the multirange are :
    On range input or as multirange method options:
    - min: minimal value of the range. Default: 0.
    - max: maximal value of the range. Default: 100.
    - step: precision of the range. Default: 1.
    - currency: symbol preceding the displayed values. Default: Empty.
    - currencyPosition: currency before/after value. Default: "after".
    - value: the current value of the range. Default: "0,100".

    As multirange method options only:
    - displayCounterInput: if we display the value. Default: true.

Initialisation of a multiple range input can be done in two ways:

Having the inputs with the options as properties and the multiple
property set will let the library initialise it just after ContentLoad

<input type="range" multiple="multiple" class="custom-range"
 min=2 max=10 step=0.5 data-currency="€" data-currency-position="before"
 value="4,8"/>

Providing a jQuery Object containing the inputs plus an Object with the
options desired.

<input id="multi" type="range" class="custom-range"/>

multirange($('#multi'), {
    min: 2,
    max: 10,
    step: 0.5,
    currency: "€",
    currencyPosition: "before",
    value: "4,8"
});

task-2000427