[FIX] web: display group buttons in inner groups in list view
if user manually do groupby and buttons comes in inner group then opening an
inner group doesn't work and throws traceback

we intensely did that, groupbys buttons are only displayed on the first level of
groupby so no need to fetch the extra fields for inner groups but there is a
case when the user do groupby manually and select groupbys buttons field in
inner groups.

hence call _readGroupExtraFields regardless of group level.

task-1998791