[POC] base,web: add button to see ir.property values
This POC is heavily inspired by the translate fields

Currently the code is extracted into a mixin which is directly inherited
by the AbstractField class. We should decide if we want this to be done
directly in the AbstractField or if it should be implemented in each
fields that needs it

ir.property currently support these fields ([X] means that it is used in
the codebase):
  - Boolean []
  - Integer [X]
  - Float [X]
  - Date [X]
  - Datetime [X]
  - Char [X]
  - Text []
  - Selection [X]
  - Many2one [X]
  - Binary []

Currently the button is displayed both in edit and readonly modes, this
can be changed and displayed only in one of those modes.
We could also show it only when the user is in debug mode, however this
would introduce asynchronous code in the rendering of the fields which
can impact the usability. That said it would only impact the
company_dependant fields (this means 48 fields as of 12/06/2019).

The method `get_values_action` should be renamed and maybe reworked to
ensure it cannot easily break/fail.

Currently no ir.rule changes was needed, it will only show the values
for the companies you asked.

The look is based on what is done with translations, but can certainly
be improved a lot (firstly changing the icon might be necessary)

In order to display the value of the properties in the tree view, we
could override the fields_view_get and add the corresponding field when
necessary.

Task 12208