[IMP] pos: Add CUPS drivers for ESC/POS printers
To be able to print PDFs on thermal printers, we need to have the right
CUPS drivers, so we add them for all supported ESCPOS printers.

Epson's official driver hasn't been compiled for ARM architecture so
it cannot run on RPi. We found an alternative on GitHub that works with
most ESCPOS printers, so Star ones too.

TaskID: 1939455