[IMP] sale_expense: only salesman can select SO on expense
Since 12.0, we wanted the employee to select directly the sales order on which
the expense should be reinvoiced. The goal was to avoid employee to select the
analytic account (too technical and usually unknown for employee). Problem is
that we had to give read access to SO to simple employee, breaking the access
rules "See your own document".

This commit proposes another way to do that: now that employee can select
project on expense (setting the analytic account), we can only display the sales
order field for salesman and only if the expense can be reinvoiced (aka if the
product has an expense policy). Falty ir.rule and access rights are now removed.
The condition to be reinvoice stays the same: having an expense policy and an
analytic account set on the expense (through project or sales order).

Task-1843241