[IMP] website, website_sale: keep track of products viewed
Now when a user browse products in eCommerce, we keep track of the
products he looked at for more than 3 sec. If the user is connected it
is stored in the DB, else it is stored in session and as a cookie.
When a user creates an account, its products viewed are stored in DB
from the session/cookie and those are deleted.

The recently viewed products are displayed as a snippet but also
with the customize options in cart, product and products pages
of website_sale.

Products that are in cart will not be returned as recently viewed.

It is possible to add a recently viewed product to the cart directly
from the carousel, it will not redirect to the cart. If we are on the
cart page, the product is displayed in the cart.

task-1984575