[IMP] product, point_of_sale: improve product attribute value model
`product.template.attribute.value` now retain to which
`product.template.attribute.line` they belong to. This allows the inverse
relation (o2m) to work out of the box instead of relying on a compute.

This is easier to maintain, faster, better for data integrity, and a necessary
step towards allowing multiple lines for the same attribute.