[REF] stock: test_quant.py without demo data
skipped tests without demo data:
- test_increase_available_quantity_3
- test_decrease_available_quantity_3

These tests simulate a concurrent transaction by openning a new
cursor and lock a product we want to use with another cursor. For
this to work, a product has to be created by another of these two
cursors. Until we find a workaround, disable thest two tests.

task-2024202