[IMP] product, website_sale: speed up loading of attributes on shop
By storing which attributes are set on products, we can simplify the search for
attributes to display on the shop page.

This reduces the number of queries by 3 and with our demo data this reduces the
time of this specific line of code from around 1% of the total request down to
around 0.33%.

On the whole page load the impact is not huge but it's welcome regardless.