[IMP] mail: easy access to record from follower invitation mail
Previously, when partner is added as a follower to a record, partner is notified
via mail but there was not any link to access that record easily.

So if the invited partner is linked with user then added a button in
invitation mail for easy access of record. if partner is not linked with
any user then just send the simple mail(without access button) as it is.

task-1967234
Closes: #33270