[IMP] website, website_sale: speed up loading of categories on shop
Simplify the search for categories to display on the shop page by using
the existing relation between product and categories but defining it also on
categories.

Speed up all queries involving `child_of` and `parent_of` by using
´_parent_store=True` and `parent_path`.

Speed up `name_get` by also taking advantage of `parent_path`.

Reduce the number of queries by prefetching `child_id` for all categories.

The opportunity is also taken to improve the "no results" message to add on
which category the search was done.

Overall the impact on performance of this PR is dependent on the number of
products and categories, and it can vary from saving few queries and few ms to
saving hundred of queries and hundred of ms.
For 300 categories around 90 queries and 60ms are saved on the /shop page.

Note that while this improves the situation, the unfiltered /shop page is still
suffering when many categories are present, especially nested categories. The
impact is around 1 ms per category to give a general idea.