[IMP] website_sale: only show sub-categories for current category
Avoid loading and displaying all the sub-categories for performance. It also
makes sense functionally to show sub-categories only when the user has shown
interest in a particular top-level category.

This significantly increases the speed of the shop page when there are many
sub-categories, possibly dividing the time of the whole page by two if there
are hundred of sub-categories.