[IMP] product: changed barcode from related to compute
This commit is to make sure that even when a product is archived,
its barcode is still visible.

Since barcode is a related field to product.product, the field gets
invisible in product.template when you archive the same record from product.product.

So, we changed it to compute field.

Task-1978559